requirements.txt
# Core
Django[argon2]==5.2.4
psycopg[binary,pool]==3.2.9

# Authentication
argon2-cffi==23.1.0
django-cors-headers==4.7.0

# Langchain
langchain==0.3.26
langchain-community==0.3.27

# REST API
fastapi==0.116.1
uvicorn==0.35.0

# GraphQL
graphene==3.4.3
graphene-django==3.2.3

# Ollama
ollama==0.5.1

